lambda-sam/modules_layer/requirements.txt
pyjwt
pyjwt[crypto]
PyMySQL